cleanup and split test
