+++ src/conditional_requirement.rs
@@ -1,5 +1,4 @@
-use crate::internal::id::ConditionId;
-use crate::{Requirement, VersionSetId};
+use crate::{Requirement, VersionSetId, VersionSetUnionId, internal::id::ConditionId};
 
 /// A [`ConditionalRequirement`] is a requirement that is only enforced when a
 /// certain condition holds.
@@ -7,6 +6,7 @@ use crate::{Requirement, VersionSetId};
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct ConditionalRequirement {
     /// The requirement is enforced only when the condition evaluates to true.
+    #[cfg_attr(feature = "serde", serde(skip_serializing_if = "Option::is_none"))]
     pub condition: Option<ConditionId>,
 
     /// A requirement on another package.
@@ -17,6 +17,7 @@ pub struct ConditionalRequirement {
 /// based on whether one or more other requirements are true or false.
 #[derive(Clone, Debug, Hash, Eq, PartialEq, Ord, PartialOrd)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[cfg_attr(feature = "serde", serde(rename_all = "snake_case"))]
 pub enum Condition {
     /// Defines a combination of conditions using logical operators.
     Binary(LogicalOperator, ConditionId, ConditionId),
@@ -25,9 +26,11 @@ pub enum Condition {
     Requirement(VersionSetId),
 }
 
-/// A [`LogicalOperator`] defines how multiple conditions are compared to each other.
+/// A [`LogicalOperator`] defines how multiple conditions are compared to each
+/// other.
 #[derive(Clone, Copy, Debug, Hash, Eq, PartialEq, Ord, PartialOrd)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[cfg_attr(feature = "serde", serde(rename_all = "snake_case"))]
 pub enum LogicalOperator {
     /// The condition is true if both operands are true.
     And,
@@ -44,5 +47,17 @@ impl From<Requirement> for ConditionalRequirement {
             condition: None,
             requirement: value,
         }
+    }
+}
+
+impl From<VersionSetId> for ConditionalRequirement {
+    fn from(value: VersionSetId) -> Self {
+        Requirement::Single(value).into()
+    }
+}
+
+impl From<VersionSetUnionId> for ConditionalRequirement {
+    fn from(value: VersionSetUnionId) -> Self {
+        Requirement::Union(value).into()
     }
 }
